[Entrypoints] Reduce entrypoint complexity
